+++ class/offer.class.php
@@ -293,15 +293,15 @@ class EmergencyHouseOffer extends EmergencyHouseCommonObject
 	{
 		global $langs;
 		$labels = array(
-			self::STATUS_DRAFT => array('StatusDraft', 0),
-			self::STATUS_PENDING => array('StatusPendingValidation', 1),
-			self::STATUS_PUBLISHED => array('StatusPublished', 4),
-			self::STATUS_SUSPENDED => array('StatusSuspended', 5),
-			self::STATUS_EXPIRED => array('StatusExpired', 6),
-			self::STATUS_CLOSED => array('StatusClosed', 6),
-			self::STATUS_REJECTED => array('StatusRejected', 8),
+			self::STATUS_DRAFT => array('StatusDraft', 'status0'),
+			self::STATUS_PENDING => array('StatusPendingValidation', 'status1'),
+			self::STATUS_PUBLISHED => array('StatusPublished', 'status4'),
+			self::STATUS_SUSPENDED => array('StatusSuspended', 'status5'),
+			self::STATUS_EXPIRED => array('StatusExpired', 'status6'),
+			self::STATUS_CLOSED => array('StatusClosed', 'status6'),
+			self::STATUS_REJECTED => array('StatusRejected', 'status8'),
 		);
-		$definition = isset($labels[$status]) ? $labels[$status] : array('StatusUnknown', 0);
+		$definition = isset($labels[$status]) ? $labels[$status] : array('StatusUnknown', 'status0');
 		return dolGetStatus($langs->trans($definition[0]), $langs->trans($definition[0]), '', $definition[1], $mode);
 	}
 }